Break out utilities into files, updated webpack resolvers, move firebase into api dir
