+++ src/shared/api/firebase/initializeFirebase.js
