+++ src/shared/api/firebase/models/Auth.js
