+++ src/shared/utilities/getIn.js
@@ -1,0 +1,5 @@
+// Null safe method to retrive nested object property values
+export default (nestedObj, pathArr) => {
+  return pathArr.reduce((obj, key) =>
+    (obj && obj[key] !== 'undefined') ? obj[key] : undefined, nestedObj)
+}
