+++ src/shared/utilities/index.js
@@ -1,0 +1,4 @@
+export getIn from './getIn'
+export setIn from './setIn'
+export normalizeArray from './normalizeArray'
+export * as regex from './regex'
