+++ src/shared/utilities/regex.js
@@ -1,0 +1,4 @@
+// http://emailregex.com/
+export const EMAIL = /^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/
+
+export const USERNAME = /^[a-zA-Z0-9_-]*$/
